pyproject.toml
[project]
name = "aggregator-api"
version = "0.0.0"
description = "devs.DC"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = []

[dependency-groups]
dev = [
    "boto3==1.34.91",
    "ruff==0.2.1",
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "pytest-django==4.8.0",
    "coveralls==3.3.1",
    "tomlkit==0.12.4",
    "respx==0.21.1",
    "pytest-asyncio==0.23.6",
    "moto[dynamodb]==5.0.5",
    "mypy-boto3-dynamodb==1.34.91",
    "pytest-ruff==0.3.1",
    "pre-commit==3.7.0",
    "uvicorn==0.30.6",
    "djhtml==3.0.6",
    "beautifulsoup4==4.13.4",
    "dc-response-builder",
]

[tool.uv.sources]
aws-wsgi = { git = "https://github.com/DemocracyClub/awsgi.git", rev = "a5b800373c326e75f705d400c48f7df301409724" }
dc-design-system = { git = "https://github.com/DemocracyClub/design-system.git", tag = "0.5.0" }
dc-django-utils = { git = "https://github.com/DemocracyClub/dc_django_utils.git", tag = "8.0.3" }
dc-response-builder = { git = "https://github.com/DemocracyClub/dc_response_builder.git", tag = "1.1.2" }
dc-api-common = { workspace = true }


[tool.uv]
required-version = "==0.8.*"
package = false

[tool.uv.workspace]
members = [
    "common",
    "api/*",
]

[tool.ruff]
line-length = 80
lint.ignore = ["E501"]
lint.extend-select = [
    "I",
    "C4",
    "SIM",
    "Q003",
    "RET",
]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "@abstractmethod"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "frontend.settings.testing"
asyncio_mode = "strict"
pythonpath = [
    "api/",
    "api/frontend",
    "api/endpoints/v1",
    "api/endpoints/v1/elections",
    "api/endpoints/v1/layers_of_state",
    "api/endpoints/v1/sandbox",
    "api/endpoints/v1/voting_information",
]
